0.0.3 added webhooks

+++ bot.sh
@@ -2,8 +2,12 @@
 
 VENV_DIR=".venv"
 PYTHON_BIN="$VENV_DIR/bin/python3"
-SCRIPT="bot.py"
+GUNICORN_BIN="$VENV_DIR/bin/gunicorn"
+SCRIPT="bot:app"
 PID_FILE=".pid"
+CERT_DIR="certs"
+KEYFILE="$CERT_DIR/privkey.pem"
+CERTFILE="$CERT_DIR/fullchain.pem"
 
 start_bot() {
     if [ ! -d "$VENV_DIR" ]; then
@@ -16,10 +20,15 @@ start_bot() {
         exit 1
     fi
 
-    echo "Starting bot..."
-    nohup "$PYTHON_BIN" "$SCRIPT" > bot.out 2>&1 &  # Запуск в фоновом режиме с nohup
-    BOT_PID=$!  # Получаем PID последнего запущенного процесса
-    echo $BOT_PID > "$PID_FILE"  # Сохраняем PID в файл
+    if [ ! -f "$KEYFILE" ] || [ ! -f "$CERTFILE" ]; then
+        echo "SSL certificate files not found: $KEYFILE, $CERTFILE"
+        exit 1
+    fi
+
+    echo "Starting bot with Gunicorn..."
+    nohup "$GUNICORN_BIN" "$SCRIPT" -k uvicorn.workers.UvicornWorker --keyfile "$KEYFILE" --certfile "$CERTFILE" --bind 0.0.0.0:8443 > bot.out 2>&1 &
+    BOT_PID=$!
+    echo $BOT_PID > "$PID_FILE"
     echo "Bot started with PID $BOT_PID."
 }
